refactor: extracted generic functions
will be used by the update logic
