+++ internal/system/runtime.go
@@ -1,9 +1,5 @@
 package system
 
-import "github.com/SecretSheppy/armorpaint-cloud-content-manager/internal/logger"
-
-var log = logger.Get()
-
 func Runtime(args []string) {
 	if len(args) < 3 {
 		log.Panic("Not enough arguments")
